Move vote tracking logic out of ethereum_events module
