+++ apps/src/lib/node/ledger/protocol/transactions/mod.rs
@@ -8,3 +8,9 @@
 pub(super) mod ethereum_events;
 #[cfg(not(feature = "abcipp"))]
 mod votes;
+
+#[cfg(not(feature = "abcipp"))]
+mod read;
+
+#[cfg(not(feature = "abcipp"))]
+mod update;
