+++ apps/src/lib/node/ledger/protocol/transactions/votes.rs
@@ -1,11 +1,17 @@
 //! Logic and data types relating to tracking validators' votes for pieces of
 //! data stored in the ledger, where those pieces of data should only be acted
 //! on once they have received enough votes
-use std::collections::BTreeSet;
+use std::collections::{BTreeSet, HashMap};
 
 use borsh::{BorshDeserialize, BorshSchema, BorshSerialize};
+use eyre::{eyre, Result};
+use namada::ledger::eth_bridge::storage::vote_tracked;
+use namada::ledger::storage::{DBIter, Storage, StorageHasher, DB};
 use namada::types::address::Address;
+use namada::types::storage::BlockHeight;
 use namada::types::voting_power::FractionalVotingPower;
+
+use crate::node::ledger::protocol::transactions::read;
 
 #[derive(
     Clone, Debug, PartialEq, Eq, BorshSerialize, BorshDeserialize, BorshSchema,
@@ -25,4 +31,65 @@ pub struct VoteTracking {
     // TODO: this field is redundant - we can derive whether an event is seen
     // or not from looking at `voting_power`
     pub seen: bool,
+}
+
+pub fn calculate_new_vote_tracking(
+    seen_by: &BTreeSet<(Address, BlockHeight)>,
+    voting_powers: &HashMap<(Address, BlockHeight), FractionalVotingPower>,
+) -> Result<VoteTracking> {
+    let mut seen_by_voting_power = FractionalVotingPower::default();
+    for (validator, block_height) in seen_by {
+        match voting_powers
+            .get(&(validator.to_owned(), block_height.to_owned()))
+        {
+            Some(voting_power) => seen_by_voting_power += voting_power,
+            None => {
+                return Err(eyre!(
+                    "voting power was not provided for validator {}",
+                    validator
+                ));
+            }
+        };
+    }
+
+    let newly_confirmed =
+        seen_by_voting_power > FractionalVotingPower::TWO_THIRDS;
+    Ok(VoteTracking {
+        voting_power: seen_by_voting_power,
+        seen_by: seen_by
+            .iter()
+            .map(|(validator, _)| validator.to_owned())
+            .collect(),
+        seen: newly_confirmed,
+    })
+}
+
+pub fn calculate_updated_vote_tracking<D, H, T>(
+    store: &mut Storage<D, H>,
+    keys: &vote_tracked::Keys<T>,
+    _voting_powers: &HashMap<(Address, BlockHeight), FractionalVotingPower>,
+) -> Result<VoteTracking>
+where
+    D: 'static + DB + for<'iter> DBIter<'iter> + Sync,
+    H: 'static + StorageHasher + Sync,
+    T: BorshDeserialize,
+{
+    let _body: T = read::value(store, &keys.body())?;
+    let seen: bool = read::value(store, &keys.seen())?;
+    let seen_by: BTreeSet<Address> = read::value(store, &keys.seen_by())?;
+    let voting_power: FractionalVotingPower =
+        read::value(store, &keys.voting_power())?;
+
+    let vote_tracking = VoteTracking {
+        voting_power,
+        seen_by,
+        seen,
+    };
+
+    tracing::warn!(
+        ?vote_tracking,
+        "Updating events is not implemented yet, so the returned VoteTracking \
+         will be identical to the one in storage",
+    );
+    Ok(vote_tracking)
 }
